move views to separate files
